requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
beartype==0.16.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==1.2.2
cmaes==0.9.1
colorlog==6.7.0
cycler==0.11.0
Cython==0.29.35
datasets==2.13.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
dm-control==1.0.12
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.7.0
ema-pytorch==0.2.5
entrypoints==0.4
et-xmlfile==1.1.0
fasteners==0.18
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
greenlet==2.0.2
grpcio==1.56.2
gym==0.23.0
gym-notices==0.0.8
h5py==3.8.0
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
labmaze==1.0.6
lightning-utilities==0.8.0
lxml==4.9.2
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
mujoco==2.3.5
mujoco-py==2.1.2.14
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.21.6
omegaconf==2.3.0
openpyxl==3.1.2
optuna==3.2.0
packaging==23.1
pandas==1.3.5
parmap==1.6.0
parso==0.8.3
patchelf==0.17.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pyarrow==12.0.1
pybullet==3.2.5
pycparser==2.21
pygame==2.4.0
pyglet==1.3.2
Pygments==2.15.1
PyOpenGL==3.1.7
pyparsing==3.0.9
pyrallis==0.3.1
pyrootutils==1.0.4
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytorch-lightning==1.9.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
ray==2.6.2
regex==2023.6.3
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.16
sympy==1.12
tensordict==0.1.2
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.1.0
torchdiffeq==0.2.3
torchmetrics==1.2.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wandb==0.15.4
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
